pyproject.toml
[project]
name = "mcp-excel-server"
version = "1.0.0"
description = "A Message Control Protocol (MCP) server for Excel"
readme = "README.md"

dependencies = [
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "mcp>=1.9.2",
    "openpyxl>=3.1.5",
    "typer>=0.16.0",
    "uvx<2.0",
    "mcp-server>=0.1.4",
    "pandas>=2.3.0",
    "pytest>=8.3.5",
    "ruff >=0.7.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_excel_server"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"


[tool.hatch.build]
packages = ["src/mcp_excel_server"]
